+++ NEWS.md
@@ -10,6 +10,7 @@
 - Stem segments fixed to two
 - functional parameter pRootDisc removed from Sperry's advanced model
 - Bug correction in fuel_cohortFineFMC
+- Remove ksympver and add klatleaf/klatstem to control parameters
 
 # Version 0.8.3
 - Reference book (medfatebook) launched